chore(gateway): Deduplicate common schemas
